Update val.prob.ci.2
Logistic calibration curve: fix for predicted probabilities 0 or 1
Loess CI BT: CI's of the Loess curve derived by bootstrapping can now also be filled (something Nieboer Daan noticed when he was testing the    code)
RCS: in some cases, there were estimation problems with 5 knots. If the latter would occur, RCS is rerun with 4 knots and with 3 knots if estimation problems persist with 4 knots.
Stats: statistics will again be added to the plot, as in the original function val.prob.ci. However, the placement of the statistics and the legend is not optimal. The user can specify the location of both the statistics and the legend in statloc and legendloc, respectively